refactor: move help document catalog
