+++ src/app/lib.rs
@@ -1,3 +1,4 @@
+pub mod catalog;
 pub mod cmd;
 pub mod model;
 pub mod policy;
